Added <postcheck> support to the test harness.

+++ tests/runtests.pl
@@ -19,7 +19,7 @@
 # This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
 # KIND, either express or implied.
 #
-# $Id: runtests.pl,v 1.234 2007-04-10 00:37:56 danf Exp $
+# $Id: runtests.pl,v 1.235 2007-04-23 22:58:45 danf Exp $
 ###########################################################################
 # These should be the only variables that might be needed to get edited:
 
@@ -1633,6 +1633,7 @@ sub singletest {
     my $dumped_core;
     my $cmdres;
 
+    # Apr 2007: precommand isn't being used and could be removed
     my @precommand= getpart("client", "precommand");
     if($precommand[0]) {
         # this is pure perl to eval!
@@ -1654,7 +1655,7 @@ sub singletest {
     }
     # run the command line we built
     if ($torture) {
-        return torture($CMDLINE,
+        $cmdres = torture($CMDLINE,
                        "$gdb --directory libtest $DBGCURL -x log/gdbcmd");
     }
     elsif($gdbthis) {
@@ -1692,12 +1693,31 @@ sub singletest {
         }
     }
 
+    # run the postcheck command
+    my @postcheck= getpart("client", "postcheck");
+    $cmd = $postcheck[0];
+    chomp $cmd;
+    subVariables \$cmd;
+    if($cmd) {
+	my $rc = system("$cmd");
+	if($rc != 0) {
+	    logmsg "postcheck failure\n";
+	    return 1;
+	}
+	logmsg "postchecked $cmd\n" if($verbose);
+    }
+
     # remove the special FTP command file after each test!
     unlink($FTPDCMD);
 
     my $e;
     for $e (@envs) {
         $ENV{$e}=""; # clean up
+    }
+
+    # Don't bother doing verification on torture tests
+    if ($torture) {
+        return $cmdres;
     }
 
     my @err = getpart("verify", "errorcode");
